uml_experiment/requirements.txt
django==5.0
mysqlclient
google-genai
pillow
django-crispy-forms
crispy-bootstrap5
pandas
matplotlib
seaborn
scipy
django-widget-tweaks
